requirements.txt
pkg-resources==0.0.0
ruamel.yaml==0.15.87
xmltodict==0.11.0
